+++ crates/sncast/src/starknet_commands/account/mod.rs
@@ -17,9 +17,8 @@ use sncast::helpers::rpc::RpcArgs;
 use sncast::response::explorer_link::block_explorer_link_if_allowed;
 use sncast::response::ui::UI;
 use sncast::{AccountType, chain_id_to_network_name, decode_chain_id};
-use sncast::{WaitForTx, get_chain_id};
+use sncast::{SignerSource, SignerType, WaitForTx, get_chain_id};
 use starknet_rust::providers::Provider;
-use starknet_rust::signers::SigningKey;
 use starknet_types_core::felt::Felt;
 use std::collections::BTreeMap;
 use std::io::{self, IsTerminal};
@@ -47,8 +46,10 @@ pub enum Commands {
     List(List),
 }
 
+#[allow(clippy::too_many_arguments)]
 pub fn prepare_account_json(
-    private_key: &SigningKey,
+    signer_type: &SignerType,
+    public_key: Felt,
     address: Felt,
     deployed: bool,
     legacy: bool,
@@ -61,14 +62,24 @@ pub fn prepare_account_json(
         AccountType::Argent => AccountType::Ready,
         _ => account_type,
     };
+
     let mut account_json = json!({
-        "private_key": format!("{:#x}", private_key.secret_scalar()),
-        "public_key": format!("{:#x}", private_key.verifying_key().scalar()),
+        "public_key": format!("{public_key:#x}"),
         "address": format!("{address:#x}"),
         "type": format!("{saved_account_type}").to_lowercase().replace("openzeppelin", "open_zeppelin"),
         "deployed": deployed,
         "legacy": legacy,
     });
+
+    match signer_type {
+        SignerType::Local { private_key } => {
+            account_json["private_key"] = serde_json::Value::String(format!("{private_key:#x}"));
+        }
+        SignerType::Ledger { ledger_path } => {
+            account_json["ledger_path"] =
+                serde_json::Value::String(ledger_path.derivation_string());
+        }
+    }
 
     if let Some(salt) = salt {
         account_json["salt"] = serde_json::Value::String(format!("{salt:#x}"));
@@ -231,6 +242,14 @@ pub async fn account(
             let provider = create.rpc.get_provider(&config, ui).await?;
 
             let chain_id = get_chain_id(&provider).await?;
+
+            let signer_type = create
+                .ledger_key_locator
+                .resolve(ui)
+                .map(|ledger_path| SignerType::Ledger { ledger_path });
+
+            let signer_source = SignerSource::new(config.keystore.clone(), signer_type.as_ref())?;
+
             let account = if config.keystore.is_none() {
                 create
                     .name
@@ -242,11 +261,11 @@ pub async fn account(
             let result = starknet_commands::account::create::create(
                 &account,
                 &config.accounts_file,
-                config.keystore.as_ref(),
                 &provider,
                 chain_id,
                 &create,
                 &config,
+                &signer_source,
                 ui,
             )
             .await;
@@ -264,7 +283,6 @@ pub async fn account(
             let fee_args = deploy.fee_args.clone();
 
             let chain_id = get_chain_id(&provider).await?;
-            let keystore_path = config.keystore.clone();
             let result = starknet_commands::account::deploy::deploy(
                 &provider,
                 &config.accounts_file,
@@ -272,7 +290,7 @@ pub async fn account(
                 chain_id,
                 wait_config,
                 &config.account,
-                keystore_path,
+                config.keystore.clone(),
                 fee_args,
                 ui,
             )
@@ -283,16 +301,17 @@ pub async fn account(
 
             if config.keystore.is_none()
                 && run_interactive_prompt
-                && let Err(_) = prompt_to_add_account_as_default(
-                    &deploy
+                && let Err(err) = prompt_to_add_account_as_default(
+                    deploy
                         .name
-                        .expect("Must be provided if not using a keystore"),
+                        .as_ref()
+                        .expect("Must be provided when using accounts file"),
                 )
             {
                 // TODO(#3436)
                 ui.print_error(
                     "account deploy",
-                    "Error: Failed to launch interactive prompt: {err}".to_string(),
+                    format!("Error: Failed to launch interactive prompt: {err}"),
                 );
             }