+++ crates/sncast/src/starknet_commands/multicall/run.rs
@@ -1,23 +1,27 @@
+use std::str::FromStr;
+
 use crate::starknet_commands::invoke::execute_calls;
+use crate::starknet_commands::multicall::contract_registry::ContractRegistry;
+use crate::starknet_commands::multicall::deploy::MulticallDeploy;
+use crate::starknet_commands::multicall::invoke::MulticallInvoke;
+use crate::starknet_commands::utils::felt_or_id::FeltOrId;
 use anyhow::{Context, Result};
 use camino::Utf8PathBuf;
 use clap::Args;
 use serde::Deserialize;
-use sncast::helpers::constants::UDC_ADDRESS;
+use serde_json::Number;
+use sncast::WaitForTx;
 use sncast::helpers::fee::FeeArgs;
 use sncast::helpers::rpc::RpcArgs;
 use sncast::response::errors::handle_starknet_command_error;
 use sncast::response::multicall::run::MulticallRunResponse;
 use sncast::response::ui::UI;
-use sncast::{WaitForTx, extract_or_generate_salt, udc_uniqueness};
-use starknet_rust::accounts::{Account, SingleOwnerAccount};
+use starknet_rust::accounts::SingleOwnerAccount;
 use starknet_rust::core::types::Call;
-use starknet_rust::core::utils::{get_selector_from_name, get_udc_deployed_address};
 use starknet_rust::providers::JsonRpcClient;
 use starknet_rust::providers::jsonrpc::HttpTransport;
-use starknet_rust::signers::LocalWallet;
+use starknet_rust::signers::Signer;
 use starknet_types_core::felt::Felt;
-use std::collections::HashMap;
 
 #[derive(Args, Debug, Clone)]
 #[command(about = "Execute a multicall from a .toml file", long_about = None)]
@@ -35,100 +39,73 @@ pub struct Run {
 
 #[derive(Deserialize, Debug)]
 #[serde(untagged)]
-enum Input {
+pub enum Input {
     String(String),
-    Number(i64),
+    Number(Number),
 }
 
 #[derive(Deserialize, Debug)]
-struct DeployCall {
-    class_hash: Felt,
-    inputs: Vec<Input>,
-    unique: bool,
-    salt: Option<Felt>,
-    id: String,
+#[serde(tag = "call_type", rename_all = "lowercase")]
+enum CallItem {
+    Deploy(DeployItem),
+    Invoke(InvokeItem),
 }
 
 #[derive(Deserialize, Debug)]
-struct InvokeCall {
-    contract_address: String,
-    function: String,
-    inputs: Vec<Input>,
+struct MulticallFile {
+    #[serde(rename = "call")]
+    calls: Vec<CallItem>,
 }
 
-pub async fn run(
+#[derive(Deserialize, Debug)]
+pub struct DeployItem {
+    pub class_hash: Felt,
+    pub inputs: Vec<Input>,
+    pub unique: bool,
+    pub salt: Option<Felt>,
+    pub id: Option<String>,
+}
+
+#[derive(Deserialize, Debug)]
+pub struct InvokeItem {
+    pub contract_address: FeltOrId,
+    pub function: String,
+    pub inputs: Vec<Input>,
+}
+
+pub async fn run<S>(
     run: Box<Run>,
-    account: &SingleOwnerAccount<&JsonRpcClient<HttpTransport>, LocalWallet>,
+    account: &SingleOwnerAccount<&JsonRpcClient<HttpTransport>, S>,
+    provider: &JsonRpcClient<HttpTransport>,
     wait_config: WaitForTx,
     ui: &UI,
-) -> Result<MulticallRunResponse> {
+) -> Result<MulticallRunResponse>
+where
+    S: Signer + Sync + Send,
+{
     let fee_args = run.fee_args.clone();
 
     let contents = std::fs::read_to_string(&run.path)?;
-    let items_map: HashMap<String, Vec<toml::Value>> =
+    let multicall: MulticallFile =
         toml::from_str(&contents).with_context(|| format!("Failed to parse {}", run.path))?;
 
-    let mut contracts = HashMap::new();
+    let mut contracts = ContractRegistry::new(provider);
     let mut parsed_calls: Vec<Call> = vec![];
 
-    for call in items_map.get("call").unwrap_or(&vec![]) {
-        let call_type = call.get("call_type");
-        if call_type.is_none() {
-            anyhow::bail!("`Field call_type` is missing in a call specification");
-        }
-
-        match call_type.unwrap().as_str() {
-            Some("deploy") => {
-                let deploy_call: DeployCall = toml::from_str(toml::to_string(&call)?.as_str())
-                    .context("Failed to parse toml `deploy` call")?;
-
-                let salt = extract_or_generate_salt(deploy_call.salt);
-                let mut calldata = vec![
-                    deploy_call.class_hash,
-                    salt,
-                    Felt::from(u8::from(deploy_call.unique)),
-                    deploy_call.inputs.len().into(),
-                ];
-
-                let parsed_inputs = parse_inputs(&deploy_call.inputs, &contracts)?;
-                calldata.extend(&parsed_inputs);
-
-                parsed_calls.push(Call {
-                    to: UDC_ADDRESS,
-                    selector: get_selector_from_name("deployContract")?,
-                    calldata,
-                });
-
-                let contract_address = get_udc_deployed_address(
-                    salt,
-                    deploy_call.class_hash,
-                    &udc_uniqueness(deploy_call.unique, account.address()),
-                    &parsed_inputs,
-                );
-                contracts.insert(deploy_call.id, contract_address.to_string());
+    for call in multicall.calls {
+        match call {
+            CallItem::Deploy(item) => {
+                let call = MulticallDeploy::new_from_item(&item, &contracts)?
+                    .build_call(account, &mut contracts)
+                    .await?;
+                parsed_calls.push(call);
             }
-            Some("invoke") => {
-                let invoke_call: InvokeCall = toml::from_str(toml::to_string(&call)?.as_str())
-                    .context("Failed to parse toml `invoke` call")?;
-                let mut contract_address = &invoke_call.contract_address;
-                if let Some(addr) = contracts.get(&invoke_call.contract_address) {
-                    contract_address = addr;
-                }
-
-                let calldata = parse_inputs(&invoke_call.inputs, &contracts)?;
-
-                parsed_calls.push(Call {
-                    to: contract_address
-                        .parse()
-                        .context("Failed to parse contract address to Felt")?,
-                    selector: get_selector_from_name(&invoke_call.function)?,
-                    calldata,
-                });
+            CallItem::Invoke(item) => {
+                let call = MulticallInvoke::new_from_item(&item, &contracts)?
+                    .build_call(&mut contracts)
+                    .await?;
+                parsed_calls.push(call);
             }
-            Some(unsupported) => {
-                anyhow::bail!("Unsupported call type found = {unsupported}");
-            }
-            None => anyhow::bail!("Field `call_type` is missing in a call specification"),
         }
     }
 
@@ -138,17 +115,15 @@ pub async fn run(
         .map_err(handle_starknet_command_error)
 }
 
-fn parse_inputs(inputs: &Vec<Input>, contracts: &HashMap<String, String>) -> Result<Vec<Felt>> {
+pub fn parse_inputs(inputs: &[Input], contract_registry: &ContractRegistry) -> Result<Vec<Felt>> {
     let mut parsed_inputs = Vec::new();
     for input in inputs {
         let felt_value = match input {
-            Input::String(s) => {
-                let resolved = contracts.get(s).unwrap_or(s);
-                resolved
-                    .parse()
-                    .context(format!("Failed to parse input '{resolved}' to Felt"))?
-            }
-            Input::Number(n) => (*n).into(),
+            Input::String(s) => contract_registry
+                .get_address_by_id(s)
+                .map_or_else(|| s.parse(), Ok)?,
+            Input::Number(n) => Felt::from_str(&n.to_string())
+                .with_context(|| format!("Failed to parse {n} to felt"))?,
         };
         parsed_inputs.push(felt_value);
     }