+++ crates/sncast/tests/helpers/fixtures.rs
@@ -210,7 +210,7 @@ pub async fn invoke_contract(
         url: Some(Url::parse(URL).expect("Failed to parse URL")),
         network: None,
     };
-    let account = get_account(&config, &provider, &rpc_args, None, &UI::default())
+    let account = get_account(&config, &provider, &rpc_args, &UI::default())
         .await
         .expect("Could not get the account");
 
@@ -228,6 +228,11 @@ pub async fn invoke_contract(
         selector: get_selector_from_name(entry_point_name)
             .unwrap_or_else(|_| panic!("Could not get selector from {entry_point_name}")),
         calldata,
+    };
+
+    let account = match account {
+        sncast::AccountVariant::LocalWallet(acc) => acc,
+        sncast::AccountVariant::Ledger(_) => panic!("Ledger account not supported in test"),
     };
 
     let execution = account.execute_v3(vec![call]);